Separate Moodle functions into modules
